requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.1
pyasn1-modules==0.3.0
pyasn1==0.5.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
rsa==4.9
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.6
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1